pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecf-agent"
version = "1.0.0"
description = "Agente de recolección de comprobantes fiscales electrónicos (e-CF)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "TekServices", email = "dev@tekservices.com"}
]
keywords = ["ecf", "dgii", "facturación", "dominican-republic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pymysql>=1.1",
    "psycopg2-binary>=2.9",
    "pyodbc>=5.0",
    "jsonata>=0.4",
    "httpx>=0.25",
    "apscheduler>=3.10",
    "loguru>=0.7",
    "click>=8.1",
]

[project.optional-dependencies]
oracle = ["cx_Oracle>=8.3"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.5",
]

[project.scripts]
ecf-agent = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
